DatabaseWrapper + database package refactor
