+++ smrt-server-analysis/src/test/scala/JobManagerServiceSpec.scala
@@ -8,6 +8,7 @@ import com.pacbio.secondary.analysis.configloaders.{EngineCoreConfigLoader, Pbsm
 import com.pacbio.secondary.analysis.jobtypes.SimpleDevJobOptions
 import com.pacbio.secondary.smrtlink.JobServiceConstants
 import com.pacbio.secondary.smrtlink.actors._
+import com.pacbio.secondary.smrtlink.database.Dal
 import com.pacbio.secondary.smrtlink.app.SmrtLinkConfigProvider
 import com.pacbio.secondary.smrtlink.services.{JobRunnerProvider, JobManagerServiceProvider}
 import com.pacbio.secondary.smrtlink.tools.SetupMockData
